Added separate test file and made expressor test fully check the structure of monomials; added function Monomial.getUniqueStructureString() for this purpose.
